Metodos en backend para determinar semillas en Congruencia Mixta.

+++ database/migrations/2025_05_29_013002_create_semillas_pruebas.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('v1');
             $table->integer('v2');
-            $table->integer('m');
+            $table->bigInteger('m');
             $table->string('metodo');
             $table->timestamps();
         });
